requirements.txt
fastapi
pydantic
uvicorn
SQLAlchemy==2.0.40
psycopg2-binary
alembic==1.15.2
passlib==1.7.4
python-jose==3.4.0
python-multipart
black
python-dotenv
pytest
pytest-asyncio
httpx
pydantic-settings
